refactor(car): consolidate remaining mutation rollback into the mutation-options
carAddMutationOptions and carRemoveMutationOptions each define their own
mutationKey and onError, moving already-working rollback logic out of
use-add.ts/use-delete.ts. Those hooks, plus use-edit.ts, now compose over
the base onError (call it, then toast) instead of re-deriving the rollback.

use-edit.ts's two invalidateQueries calls (byId + infinite) collapse into
one on queryKeys.all().
