+++ src/module/car/presentation/tanstack/mutation-options/remove.ts
@@ -2,7 +2,9 @@ import type { QueryClient } from '@tanstack/react-query';
 import { mutationOptions } from '@tanstack/react-query';
 
 import { carApiClient } from '@/car/dependency/api-client';
+import { CARS_INFINITE_QUERY_PAGE_DATA_LIMIT } from '@/car/infrastructure/data-source/car';
 import {
+  addCarToInfiniteQueryData,
   type CarsInfiniteQueryData,
   deepCopyCarsInfiniteQueryData,
   deleteCarFromInfiniteQueryData,
@@ -13,6 +15,7 @@ import { queryKeys } from '@/car/presentation/tanstack/query/keys';
 export const carRemoveMutationOptions = (queryClient: QueryClient) =>
   mutationOptions({
     throwOnError: false,
+    mutationKey: queryKeys.infinite(),
     mutationFn: async (carId: string) => {
       const removeResult = await carApiClient.remove(carId);
 
@@ -49,5 +52,33 @@ export const carRemoveMutationOptions = (queryClient: QueryClient) =>
       );
 
       return deletedCarContext;
+    },
+    onError: (_error, _variables, context) => {
+      if (
+        !context ||
+        context.deletedCar === null ||
+        context.deletedCarPageIndex == null ||
+        context.deletedCarPagePositionIndex == null
+      ) {
+        return;
+      }
+
+      const previousData = queryClient.getQueryData<CarsInfiniteQueryData>(
+        queryKeys.infinite(),
+      );
+
+      if (!previousData) return;
+
+      const updatedQueryData = deepCopyCarsInfiniteQueryData(previousData);
+
+      addCarToInfiniteQueryData(
+        context.deletedCar,
+        updatedQueryData,
+        CARS_INFINITE_QUERY_PAGE_DATA_LIMIT,
+        context.deletedCarPageIndex,
+        context.deletedCarPagePositionIndex,
+      );
+
+      queryClient.setQueryData(queryKeys.infinite(), updatedQueryData);
     },
   });
